Add support of overloaded methods through the annotation `@JSOverloadName`
This will allow the java code to define the annotation on overloaded methods.

This annotation added on a method changes how the name will be generated. This is used when having multiple overload of the same method to provide clear and concise name based on the given parameters. If no value is provided to the annotation, the generated name is as follow:

```java
@JSOverloadName
public void overloadMethod(String firstParam, int secondParam, CustomClass thirdParam) {
}
```
```js
prototype.overloadMethod$String_int_CustomClass = function(firstParam, secondParam, thirdParam) {
}
```